# crm-it-shkola-rtk/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports/interactions": {
      "get": {
        "operationId": "ReportsController_getInteractionsReport",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          },
          {
            "name": "onlyOverdue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "responsibleUserId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itProductId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itDirectionId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-09-30"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-09-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InteractionReportItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Реестр взаимодействий с фильтрами (период/вуз/направление/продукт/ответственный)",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/interactions/export": {
      "get": {
        "operationId": "ReportsController_exportInteractionsReport",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          },
          {
            "name": "onlyOverdue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "responsibleUserId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itProductId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itDirectionId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "xls",
                "xlsx",
                "pdf"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionsReportExportDto"
                }
              }
            }
          }
        },
        "summary": "Экспорт отфильтрованного реестра взаимодействий (файл в MinIO + временная ссылка)",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/charts": {
      "get": {
        "operationId": "ReportsController_getCharts",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          },
          {
            "name": "responsibleUserId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itProductId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "itDirectionId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-09-30"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-09-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Данные для графиков отчётного дашборда (pie/line/bar)",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/jobs": {
      "post": {
        "description": "Возвращает job (id = jobId). Статус и ссылка на файл — GET /reports/export-status/{jobId}.",
        "operationId": "ReportsController_createReportJob",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportJobDto"
                }
              }
            }
          }
        },
        "summary": "Поставить отчёт/экспорт в очередь (INTERACTIONS_EXPORT xls|xlsx|pdf, LICENSE_RADAR, SLA_RADAR)",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/export-status/{jobId}": {
      "get": {
        "description": "Задание видит его автор и Администратор. Готовые задания хранятся ~1 час, затем 404.",
        "operationId": "ReportsController_getExportStatus",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "42",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportJobDto"
                }
              }
            }
          }
        },
        "summary": "Статус задания отчёта; при SUCCESS — временная ссылка на файл (resultUrl)",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM ИТ Школа РТК",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InteractionReportItemDto": {
        "type": "object",
        "properties": {
          "interactionInstanceId": {
            "type": "string",
            "example": "b3000000-0000-4000-8000-000000000001"
          },
          "universityId": {
            "type": "string",
            "nullable": true,
            "example": "a5000000-0000-4000-8000-000000000001"
          },
          "universityName": {
            "type": "string",
            "nullable": true,
            "example": "СПбГУ (демо)"
          },
          "itDirectionId": {
            "type": "string",
            "nullable": true,
            "example": "a1000000-0000-4000-8000-000000000001"
          },
          "itDirectionName": {
            "type": "string",
            "nullable": true,
            "example": "Импортированные продукты (реестр вендоров)"
          },
          "itProductId": {
            "type": "string",
            "nullable": true,
            "example": "a3000000-0000-4000-8000-000000000001"
          },
          "itProductName": {
            "type": "string",
            "nullable": true,
            "example": "Базис Dynamix"
          },
          "currentStatusId": {
            "type": "string",
            "example": "b1000000-0000-4000-8000-000000000002"
          },
          "currentStatusName": {
            "type": "string",
            "example": "Согласование договора"
          },
          "currentPhase": {
            "enum": [
              "INITIATION",
              "NEGOTIATION",
              "CONTRACTING",
              "IMPLEMENTATION",
              "ACTIVE_USE",
              "RENEWAL",
              "TERMINATION"
            ],
            "type": "string",
            "example": "CONTRACTING"
          },
          "responsibleUserId": {
            "type": "string",
            "nullable": true,
            "example": "c0000000-0000-4000-8000-000000000001"
          },
          "responsibleUserName": {
            "type": "string",
            "nullable": true,
            "example": "Иванова Мария Сергеевна"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-01T09:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-20T14:30:00.000Z"
          },
          "daysInCurrentStatus": {
            "type": "number",
            "example": 12,
            "description": "Сколько дней взаимодействие находится в текущем статусе"
          },
          "isOverdue": {
            "type": "boolean",
            "example": false,
            "description": "Превышен ли SLA-порог для текущего статуса"
          },
          "needsReview": {
            "type": "boolean",
            "example": false,
            "description": "true — требует ручной проверки (вуз/направление не определены автоматически из интеграции)"
          }
        },
        "required": [
          "interactionInstanceId",
          "currentStatusId",
          "currentStatusName",
          "currentPhase",
          "createdAt",
          "updatedAt",
          "daysInCurrentStatus",
          "isOverdue",
          "needsReview"
        ]
      },
      "InteractionsReportExportDto": {
        "type": "object",
        "properties": {
          "format": {
            "enum": [
              "xls",
              "xlsx",
              "pdf"
            ],
            "type": "string",
            "example": "xlsx"
          },
          "fileName": {
            "type": "string",
            "example": "reestr-vzaimodeystviy-2026-09-26.xlsx"
          },
          "url": {
            "type": "string",
            "example": "https://minio.internal/crm-exports/reestr-vzaimodeystviy-2026-09-26.xlsx?X-Amz-Expires=600&..."
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "rowCount": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "format",
          "fileName",
          "url",
          "generatedAt",
          "rowCount"
        ]
      },
      "ChartSlicePointDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Согласование договора"
          },
          "value": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
      "ChartTimeSeriesPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-01"
          },
          "value": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "date",
          "value"
        ]
      },
      "ChartsResponseDto": {
        "type": "object",
        "properties": {
          "statusDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChartSlicePointDto"
            }
          },
          "interactionsOverTime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChartTimeSeriesPointDto"
            }
          },
          "licensesByProduct": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChartSlicePointDto"
            }
          }
        },
        "required": [
          "statusDistribution",
          "interactionsOverTime",
          "licensesByProduct"
        ]
      },
      "CreateReportJobDto": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "LICENSE_RADAR",
              "SLA_RADAR",
              "INTERACTIONS_EXPORT"
            ],
            "type": "string",
            "example": "INTERACTIONS_EXPORT"
          },
          "format": {
            "enum": [
              "xls",
              "xlsx",
              "pdf"
            ],
            "type": "string",
            "example": "xlsx",
            "description": "Только для INTERACTIONS_EXPORT (по умолчанию xlsx). Радары отдаются файлом JSON."
          },
          "from": {
            "type": "string",
            "example": "2026-09-01"
          },
          "to": {
            "type": "string",
            "example": "2026-09-30"
          },
          "universityId": {
            "type": "string"
          },
          "itDirectionId": {
            "type": "string"
          },
          "itProductId": {
            "type": "string"
          },
          "responsibleUserId": {
            "type": "string"
          },
          "onlyOverdue": {
            "type": "boolean"
          }
        },
        "required": [
          "type"
        ]
      },
      "ReportJobDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "42",
            "description": "jobId очереди BullMQ"
          },
          "type": {
            "enum": [
              "LICENSE_RADAR",
              "SLA_RADAR",
              "INTERACTIONS_EXPORT"
            ],
            "type": "string",
            "example": "INTERACTIONS_EXPORT"
          },
          "status": {
            "enum": [
              "QUEUED",
              "PROCESSING",
              "SUCCESS",
              "FAILED"
            ],
            "type": "string",
            "example": "QUEUED"
          },
          "requestedById": {
            "type": "string",
            "example": "c0000000-0000-4000-8000-000000000001"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "resultUrl": {
            "type": "string",
            "nullable": true,
            "example": "http://localhost:3900/crm-attachments/exports/....xlsx?X-Amz-Expires=600&...",
            "description": "Временная ссылка на готовый файл в хранилище (только при status=SUCCESS); при каждом GET выдаётся заново"
          },
          "fileName": {
            "type": "string",
            "nullable": true,
            "example": "reestr-vzaimodeystviy-2026-09-26.xlsx"
          },
          "rowCount": {
            "type": "number",
            "nullable": true,
            "example": 120
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Причина ошибки (только при status=FAILED)"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "requestedById",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}